reverted dateRange into EventsForMonth and corrected the util function.
