+++ src/components/elements/EventsForMonth.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import moment from 'moment';
 import { Link } from 'react-router-dom';
 import { sortByDate } from 'utils/dates';
-import { DateRange } from 'utils/dateRange'
+// import { DateRange } from 'utils/dateRange'
 
 
 const styleClasses = {
@@ -15,6 +15,16 @@ const styleClasses = {
     rounded
     text-white
   `
+}
+
+function DateRange(event) {
+  const dateStart = event.start.split(' ')[0]
+  const dateEnd = event.end.split(' ')[0]
+
+  if(dateStart === dateEnd) {
+    return <p className="truncate">{event.start.slice(0, -6) + ' - ' + event.end.substring(event.end.indexOf(' ') + 1)}</p>
+  }
+  return <p className="truncate">{event.start + ' - ' + event.end}</p>
 }
 
 function RenderEventType(data) {
